Refactor: Relocate SplashInteractor and SplashInteractorImpl
Move `SplashInteractor` and `SplashInteractorImpl` from the `splash` subpackage to the root `interactor` package within the `startup-feature` module.

Update relevant import statements in `FeatureStartupModule.kt` and `SplashViewModel.kt` to reflect the new location of these classes.
